% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_standard_dev.R
\name{approx_ratio_sd}
\alias{approx_ratio_sd}
\title{Approximate the standard deviation of the ratio of two vectors}
\usage{
approx_ratio_sd(
  numerator_mean,
  numerator_sd,
  denominator_mean,
  denominator_sd,
  covariance
)
}
\arguments{
\item{numerator_mean}{mean of the numerator}

\item{numerator_sd}{standard dev of the numerator}

\item{denominator_mean}{mean of the denominator}

\item{denominator_sd}{standard dev of the denominator}

\item{covariance}{covariance of the numerator and denominator}
}
\value{
Returns a single float: the sd estimate
}
\description{
The delta method gives us:
\deqn{
   Var(N/D) \approx \frac{\mu_N^2}{\mu_D^2}
                    \bigg[ \frac{\sigma^2_N}{\mu_N^2} +
                           \frac{\sigma^2_D}{\mu_D^2} -
                           2 \frac{Cov(N, D)}{\mu_N \mu_D}]
}
Of course it's not necessarily clear we'd have the covariance
and not the vectors, but there are indeed some interesting
cases.
}
\details{
From e.g. Kendall’s Advanced Theory of Statistics or other
places with delta method / Taylor expansion
}
\author{
Colman Humphrey
}
