% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binary_search}
\alias{binary_search}
\title{Generic binary search algo: finding the input.}
\usage{
binary_search(
  target_value,
  monotone_function,
  init_bounds = NULL,
  error_gap = 0.000001,
  max_iters = 100L
)
}
\arguments{
\item{target_value}{The value the function should achieve.}

\item{monotone_function}{The function we want to find the
relevant input for.}

\item{init_bounds}{Default NULL; supply a length-two vector if
you wish to supply your own initial bounds. Note that if your function
only works in a certain range, or is only monotonic in a certain range,
then you'll very likely need to supply your own bounds. If not given,
we'll search in both directions from zero.}

\item{error_gap}{Binary search until the gap between the
function on the input and the target value is smaller than this number.}

\item{max_iters}{How many iterations to try at most. You can supply
\code{Inf} if you really want.}
}
\value{
The input that gives \code{target_value} as output.
}
\description{
Generic binary search algo: finding the input.
}
\author{
Colman Humphrey
}
