% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_standard_dev.R
\name{bipartite_match_sd}
\alias{bipartite_match_sd}
\title{Calculates the standard deviation for a match, on the outcome}
\usage{
bipartite_match_sd(
  x_mat,
  cov_x,
  y_vector,
  match_list,
  treat_vec,
  caliper_list = gen_caliper_list(),
  propensity_list = match_propensity_list(NULL),
  weight_vec = NULL,
  use_all_controls = TRUE,
  sqrt_mahal = TRUE
)
}
\arguments{
\item{x_mat}{input/design matrix (already rank-adjusted etc)}

\item{cov_x}{the (potentially rank-adjusted) covariance of \code{x_mat}.
This means it's possible that \code{cov(x_mat)} is not equal to
\code{cov_x}; see \code{covariance_with_ranks} for more details.}

\item{y_vector}{Outcome vector (not used in match generation).}

\item{match_list}{A particular match result.}

\item{treat_vec}{Logical (or 1/0) vector, indicating treatment (or control).}

\item{caliper_list}{Optional, see \code{gen_caliper_list}. Provide
this to force matches that are close on some metric.}

\item{use_all_controls}{logical; if using bipartite matches, should we
estimate the variance using all controls possible? Default \code{TRUE}.
If \code{FALSE}, will only use controls from the match given}
}
\value{
a single float, the standard deviation (not standard error)
  for the match
}
\description{
Calculates the standard deviation for a match, on the outcome
}
\author{
Colman Humphrey
}
