% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_weighted_matches.R
\name{brier_nonbipartite_matches}
\alias{brier_nonbipartite_matches}
\title{Computes all matches, then gets the brier scores for each. Reorder by
number of sinks.}
\usage{
brier_nonbipartite_matches(
  x_mat,
  cov_x,
  weight_list,
  tolerance_list = gen_tolerance_list(),
  match_method = c("with_replacement", "optimal", "greedy"),
  n_sinks = 0L,
  caliper_list = gen_caliper_list(),
  propensity_list = match_propensity_list(NULL),
  sqrt_mahal = TRUE,
  keep_all_with_replacement = FALSE,
  silent = !interactive()
)
}
\arguments{
\item{x_mat}{input/design matrix (already rank-adjusted etc)}

\item{cov_x}{the (potentially rank-adjusted) covariance of \code{x_mat}.
This means it's possible that \code{cov(x_mat)} is not equal to
\code{cov_x}; see \code{covariance_with_ranks} for more details.}

\item{weight_list}{list of weight vectors. See `generate_random_weights` to
automatically generate a reasonable set of vectors.}

\item{tolerance_list}{See \code{gen_tolerance_list}}

\item{match_method}{This enum corresponds to the three matching methods
discussed above:
\describe{
  \item{"with_replacement"}{Finds smallest control for each treatment}
  \item{"greedy"}{Greedily generates pairs. Note that
  the order for choosing the greedy pairs is random, which is not the only
  possible solution.}
  \item{"optimal"}{Minimum total distance}
}}

\item{n_sinks}{how many potential matches to not bother with
NOTE: you can do this as a vector, but not for optimal matching.}

\item{caliper_list}{Optional, see \code{gen_caliper_list}. Provide
this to force matches that are close on some metric.}

\item{keep_all_with_replacement}{logical, default FALSE.
When nbp matching with replacement, you can in some cases form nearly
as many pairs as there are units (if using a tolerance vec, then maybe not
the lowest value pairs, since they'll have no control unit).
But this would form a contrast with all other methods, so we will cut
down to using half by default.}
}
\value{
List of matches within sink values,
 and brier scores for each.
}
\description{
Computes all matches, then gets the brier scores for each. Reorder by
number of sinks.
}
\author{
Colman Humphrey
}
