% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_eval.R
\name{brier_score}
\alias{brier_score}
\title{Takes in training/test data and a prediction
function to use, generates brier score}
\usage{
brier_score(
  train_test_list,
  match_predict_function = match_predict_xgb(),
  avg = TRUE
)
}
\arguments{
\item{train_test_list}{output from \code{predict_prepare}}

\item{match_predict_function}{function to predict treated unit,
see e.g. \code{match_predict_xgb}
(produces the default on calling) or
\code{match_predict_linear}}

\item{avg}{logical, default TRUE: should we average (or sum) the briers?}
}
\value{
result from \code{calc_brier}, length-one double
}
\description{
Takes in training/test data and a prediction
function to use, generates brier score
}
\author{
Colman Humphrey
}
