% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_eval.R
\name{brier_score_cv}
\alias{brier_score_cv}
\title{For a match, calculates brier score using cross validation}
\usage{
brier_score_cv(
  x_mat,
  match_list,
  design = "cross_all",
  num_folds = 5,
  match_predict_function = match_predict_xgb()
)
}
\arguments{
\item{x_mat}{typical input matrix}

\item{match_list}{match result}

\item{design}{see \code{predict_prepare}}

\item{num_folds}{how many CV folds to use}

\item{match_predict_function}{function to predict treated units}
}
\value{
brier score (averaged over all units)
}
\description{
For a match, calculates brier score using cross validation
}
\author{
Colman Humphrey
}
