% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_eval.R
\name{brier_score_split}
\alias{brier_score_split}
\title{For a match, calculates brier score on a test split of the data}
\usage{
brier_score_split(
  x_mat,
  match_list,
  design = "cross_all",
  train_fraction = 0.7,
  match_predict_function = match_predict_xgb()
)
}
\arguments{
\item{x_mat}{typical input matrix}

\item{match_list}{match result}

\item{design}{see \code{predict_prepare}}

\item{train_fraction}{split of data to use for training}

\item{match_predict_function}{function to predict treated units}
}
\value{
brier score
}
\description{
For a match, calculates brier score on a test split of the data
}
\author{
Colman Humphrey
}
