% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{compute_sim_result}
\alias{compute_sim_result}
\title{Takes in functions to generate simulation data, and computes
simulation results for our method and some competitors}
\usage{
compute_sim_result(
  x_generator = default_x_generator,
  treat_prob_generator,
  mean_generator,
  error_generator = default_error_generator,
  n_sink_gen = n_sink_generator(),
  match_method = "with_replacement",
  n_rows = 500L,
  n_cols = 5L,
  num_weight_vectors = 100L,
  silent = !interactive()
)
}
\arguments{
\item{x_generator}{Function that takes number of rows and columns
and produces a data matrix of that dimension.}

\item{treat_prob_generator}{Function that takes a matrix and produces
treatment probabilities.}

\item{mean_generator}{Function that takes a matrix and produces
an expected value for each row.}

\item{error_generator}{Function that accepts a number of rows and
generates an error vector (e.g. generates normal noise). Should
generate zero-mean data.}

\item{n_sink_gen}{Default \code{n_sink_generator}, and you'll
probably want to use that: this argument should be a function that
accepts a \code{treat_vec} and produces a vector of sink numbers.}

\item{n_rows}{How many rows to generate.}

\item{n_cols}{How many columns to use.}

\item{num_weight_vectors}{How many weight vectors to generate.}

\item{silent}{Default \code{!interactive()}, if you want to
suppress messages.}
}
\value{
Returns a named list:
\describe{
\item{\code{naive_est}}{Just a number: mean difference between all
treated units and all control}
\item{\code{propensity_results}}{List of lists: each with
\code{n_sinks} and the \code{est}}
\item{\code{mahal_results}}{Same as above}
\item{\code{weighted_results}}{List of lists: each with
\code{n_sinks}, the raw brier score, the permutation brier score,
and the \code{est}}
}
}
\description{
Takes in functions to generate simulation data, and computes
simulation results for our method and some competitors
}
\author{
Colman Humphrey
}
