% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliper.R
\name{create_caliper}
\alias{create_caliper}
\title{Creates caliper penalties}
\usage{
create_caliper(caliper_list, treat_vec = NULL)
}
\arguments{
\item{caliper_list}{Result of \code{gen_caliper_list}}

\item{treat_vec}{Optional; if you only want pairs between treat and control.}
}
\value{
A matrix, either square (\code{|caliper_vec| x |caliper_vec|})
  or else \code{sum(treat_vec == 1) x sum(treat_vec == 0)}.
}
\description{
Given a vector, \code{caliper_vec}, this function sets penalties
for any pairwise differences above \code{caliper_max}; either
an \code{Inf} penalty (the default), or a continuous penalty. Typical usage will
then add the resulting matrix from this function onto a distance
matrix, say from pairwise Mahalanobis.
}
\author{
Colman Humphrey
}
