% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_utils.R
\name{default_error_generator}
\alias{default_error_generator}
\title{Default function to generate mean-zero noise. Very simple.}
\usage{
default_error_generator(n_rows)
}
\arguments{
\item{n_rows}{How many rows to produce.}
}
\value{
Returns just mean 0 variance 1 normal noise.
}
\description{
Exported just so all functions have a default in
\code{generate_simulation_input}.
}
\author{
Colman Humphrey
}
