% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_utils.R
\name{default_x_generator}
\alias{default_x_generator}
\title{A default function that generates an input data matrix.}
\usage{
default_x_generator(n_rows, n_cols)
}
\arguments{
\item{n_rows}{How many rows to produce.}

\item{n_cols}{How many columns to produce.}
}
\value{
A matrix of data
}
\description{
First generates a random covariance, then generates
normal data with that covariance (actually correlation).
}
\author{
Colman Humphrey
}
