% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_utils.R
\name{example_mean_generator}
\alias{example_mean_generator}
\title{An example function that generates a mean vector
given an input matrix.}
\usage{
example_mean_generator(x_mat)
}
\arguments{
\item{x_mat}{Numeric matrix.}
}
\value{
A vector of means, with length equal to \code{nrow(x_mat)}.
}
\description{
An example function that generates a mean vector
given an input matrix.
}
\author{
Colman Humphrey
}
