% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonbipartite_matches.R
\name{fix_nbp_match}
\alias{fix_nbp_match}
\title{Converting nbp output to common output for this package.}
\usage{
fix_nbp_match(nbp_match, nrow_match, tolerance_vec = NULL)
}
\arguments{
\item{nbp_match}{Result from nbpMatching::nonbimatch.}

\item{nrow_match}{How many real rows were there?
You feed sinks to nbpmatching sort of manually, so
this is one way to remove them.}

\item{tolerance_vec}{If this is given, we sort the matches by high / low.}
}
\description{
Takes result of NBP matching and deletes sinks,
and I guess infinite distances; finally also returns only some columns.
}
\keyword{internal}
