% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fold_indexing}
\alias{fold_indexing}
\title{Produces a list of indices split into k-folds}
\usage{
fold_indexing(index_length, num_folds)
}
\arguments{
\item{index_length}{length of vector to index}

\item{num_folds}{how many folds you need}
}
\value{
list of vectors of indices
}
\description{
Produces a list of indices split into k-folds
}
\author{
Colman Humphrey
}
