% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_standard_dev.R
\name{gen_bipartite_repeated_variance}
\alias{gen_bipartite_repeated_variance}
\title{Abadie, Alberto and Guido W Imbens (2006).
“Large sample properties of matching estimators
for average treatment effects”.
Econometrica 74.1, pp. 235–267. - see page 251.
More info at “Matching on the estimated propensity score”.
Econometrica 84.2, pp. 781–807.}
\usage{
gen_bipartite_repeated_variance(
  x_mat,
  cov_x,
  y_vector,
  control_index,
  treat_vec,
  caliper_list = gen_caliper_list(),
  weight_vec = NULL,
  use_all_controls = TRUE,
  sqrt_mahal = TRUE
)
}
\arguments{
\item{x_mat}{input/design matrix (already rank-adjusted etc)}

\item{cov_x}{the (potentially rank-adjusted) covariance of \code{x_mat}.
This means it's possible that \code{cov(x_mat)} is not equal to
\code{cov_x}; see \code{covariance_with_ranks} for more details.}

\item{y_vector}{Outcome vector (not used in match generation).}

\item{treat_vec}{Logical (or 1/0) vector, indicating treatment (or control).}

\item{caliper_list}{Optional, see \code{gen_caliper_list}. Provide
this to force matches that are close on some metric.}

\item{use_all_controls}{logical; if using bipartite matches, should we
estimate the variance using all controls possible? Default \code{TRUE}.
If \code{FALSE}, will only use controls from the match given}
}
\description{
Abadie, Alberto and Guido W Imbens (2006).
“Large sample properties of matching estimators
for average treatment effects”.
Econometrica 74.1, pp. 235–267. - see page 251.
More info at “Matching on the estimated propensity score”.
Econometrica 84.2, pp. 781–807.
}
\keyword{internal}
