% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliper.R
\name{gen_caliper_list}
\alias{gen_caliper_list}
\title{Simple wrapper to unify caliper input}
\usage{
gen_caliper_list(caliper_vec = NULL, caliper_max = NULL, continuous_mult = 100)
}
\arguments{
\item{caliper_vec}{Default NULL; numeric vector that "blocks"
matches if they're too (further than \code{caliper_max}) on this value}

\item{caliper_max}{The maximum allowed difference (exactly this difference
is allowed).}

\item{continuous_mult}{The value to multiply differences above caliper max.
Set as \code{Inf} to have infinite penalties, i.e. block matches above
the max.}
}
\value{
Either \code{NULL}, or a list with the same names as the input, after checking
  values.
}
\description{
Simple wrapper to unify caliper input
}
\author{
Colman Humphrey
}
