% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonbipartite_matches.R
\name{gen_tolerance_list}
\alias{gen_tolerance_list}
\title{Simple wrapper to unify tolerance input}
\usage{
gen_tolerance_list(
  tolerance_vec = NULL,
  tolerance_min = NULL,
  tolerance_max = NULL
)
}
\arguments{
\item{tolerance_vec}{Default NULL; numeric "continuous treatment"
vector that we use to form
non-bipartite matches: units i and j can be matched if
\eqn{\mid \texttt{tolerance_vec}[i] - \texttt{tolerance_vec}[j]\mid >
\texttt{tolerance_min}}.}

\item{tolerance_min}{See above for what this does - blocks
matches that are too close on \code{tolerance_vec}. Something like minimum
relevant difference to be a "treatment" effect. Default \code{NULL} gives
zero, i.e. only blocks equality of \code{tolerance_vec}.}

\item{tolerance_max}{Optionally we may want to also ensure
our "treatment" values aren't too far apart. E.g. we may think our
assumptions are reasonable valid for small differences
in the tolerance vector, but not for large. Or another way:
we're asking for say marginal effects: how bad is one extra beer a
day?}
}
\value{
Either \code{NULL}, or a list with the same names as the input,
  with validated values.
}
\description{
Simple wrapper to unify tolerance input
}
\author{
Colman Humphrey
}
