% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_prepare.R
\name{generate_k_fold_index}
\alias{generate_k_fold_index}
\title{Constructs a k-fold list of \code{index_list} objects for a given
\code{match_list}}
\usage{
generate_k_fold_index(match_list, num_folds = 5L)
}
\arguments{
\item{match_list}{typical \code{match_list} entry}

\item{num_folds}{how many folds you want, default 5}
}
\description{
Constructs a k-fold list of \code{index_list} objects for a given
\code{match_list}
}
\author{
Colman Humphrey
}
