% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_plots.R
\name{generate_plot_list}
\alias{generate_plot_list}
\title{Takes the result of \code{generate_p_cut_frame} and creates plot
lists based on our choise of RMSE function.}
\usage{
generate_plot_list(
  full_plot_frame,
  rmse_func = rmse_from_one_func,
  silent = !interactive()
)
}
\arguments{
\item{full_plot_frame}{Result from \code{generate_p_cut_frame}.}

\item{rmse_func}{Function that takes in a vector of estimates and computes
an RMSE.}

\item{silent}{If you don't want to log anything.}
}
\description{
Takes the result of \code{generate_p_cut_frame} and creates plot
lists based on our choise of RMSE function.
}
\author{
Colman Humphrey
}
