% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{generate_random_weights}
\alias{generate_random_weights}
\title{Generating random weights to use in matching procedures,
each summing to one.}
\usage{
generate_random_weights(
  prior_weights,
  number_vectors,
  minimum_weights = NULL,
  hierarchical_list = NULL
)
}
\arguments{
\item{prior_weights}{Must be equal to the length of your columns -
i.e. the length of the weight vectors this function will produce.
We'll use this to generate uniform random variables from 0 to this
number. Can be the same value repeated if desired.}

\item{number_vectors}{How many weight vectors you want.}

\item{minimum_weights}{If you want to set minimums, either globally or
per column. Note that this will give this minimum, and then add random
weights on top of that.}

\item{hierarchical_list}{List per group / category of variable:
\describe{
  \item{\code{"index"}}{Vector of indices that this group corresponds with}
  \item{\code{"weight"}}{Weight for this group}
  \item{\code{"variance"}}{(Optional) how much variance you want this group
to have}
}
If you use this, we'll still combine it with
`prior_weights` and `minimum_weights`}
}
\value{
list of weight vectors
}
\description{
This function generates a vector of weights for each
column in your input data. You must set prior weights
(which can all be the same), and you must set a number of vectors.
After that you have some options: set minimum weights to make sure
that we don't have too little of any given column,
and set up hierarchical info if you have categories you
want to lump together (see below)
}
\author{
Colman Humphrey
}
