% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_simulation_input}
\alias{generate_simulation_input}
\title{Generates input needed for a simulation run}
\usage{
generate_simulation_input(
  n_rows = 500L,
  n_cols = 5L,
  x_generator = default_x_generator,
  treat_prob_generator = example_treat_prob_generator,
  mean_generator = example_mean_generator,
  error_generator = default_error_generator
)
}
\arguments{
\item{n_rows}{How many rows to generate.}

\item{n_cols}{How many columns to use.}

\item{x_generator}{Function that takes number of rows and columns
and produces a data matrix of that dimension.}

\item{treat_prob_generator}{Function that takes a matrix and produces
treatment probabilities.}

\item{mean_generator}{Function that takes a matrix and produces
an expected value for each row.}

\item{error_generator}{Function that accepts a number of rows and
generates an error vector (e.g. generates normal noise). Should
generate zero-mean data.}
}
\value{
List:
\describe{
\item{\code{x_mat}}{Data Matrix}
\item{\code{treat_vec}}{Treatment Vector}
\item{\code{y_vec}}{Output Vector}
}
}
\description{
Generates input needed for a simulation run
}
\author{
Colman Humphrey
}
