% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_least_squares.R
\name{gls_action}
\alias{gls_action}
\title{Does the actual work of \code{gls_known_relative_variance}}
\usage{
gls_action(y_vector, x_mat, rel_var_mat)
}
\description{
Does the actual work of \code{gls_known_relative_variance}
}
\keyword{internal}
