% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_least_squares.R
\name{gls_known_relative_variance}
\alias{gls_known_relative_variance}
\title{Solving GLS when variance is known up to a scale factor}
\usage{
gls_known_relative_variance(y_vector, x_mat, rel_var_mat)
}
\arguments{
\item{y_vector}{Outcome vector (in future can potentially be matrix)}

\item{x_mat}{Matrix of explanatory variables. We \strong{will not include the
intercept for you} so add a column of 1s if you want this. You can
also supply just a single vector and we'll upgrade it to a matrix
(but again we will not add an intercept). Needs to match length
of \code{y_vector}}

\item{rel_var_mat}{\eqn{\Sigma} in the equations above: the variance of the
error up to some (unknown) constant factor.}
}
\value{
List with the following elements:
\itemize{
\item \code{beta_gls}: the coefficients, \eqn{\hat{\beta}}
\item \code{beta_gls_stderr}: the standard errors for the coefficients
\item \code{var_beta_gls}: the covariance matrix for \eqn{\beta}
\item \code{sigma_sq_est}: our estimate for \eqn{\sigma^2}.
If \code{x_mat} has column names, they should hopefully pass through
to the resulting vectors/matrices.
}
}
\description{
If we know that \eqn{y \sim \mathcal{N}(X \beta, \sigma^2 \Sigma)}
where \eqn{\Sigma} is known but \eqn{\beta} and \eqn{\sigma^2} aren't,
then we can solve using nearly straight-forward generalised least squares.
We return a list of results: the coefficients (\eqn{\beta}); the standard
error for those coefficients; the whole covariance matrix for \eqn{\beta}
and finally our estimate for \eqn{\sigma^2}.
}
\author{
Colman Humphrey
}
