% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_least_squares.R
\name{gls_svd_lm}
\alias{gls_svd_lm}
\title{Not using this yet, but provides a useful alternative
to \code{gls_action}: using R's in-built \code{lm}
but is slower}
\usage{
gls_svd_lm(y_vector, x_mat, rel_var_mat)
}
\description{
Not using this yet, but provides a useful alternative
to \code{gls_action}: using R's in-built \code{lm}
but is slower
}
\keyword{internal}
