% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_prepare.R
\name{index_list_from_match}
\alias{index_list_from_match}
\title{Constructs an \code{index_list} from a \code{match_list} and a logical
index for training data}
\usage{
index_list_from_match(match_list, train_index)
}
\arguments{
\item{match_list}{see \code{bipartite_matches} etc}

\item{train_index}{logical index, same length as
\code{match_list[["treat_index"]]} (and
\code{match_list[["control_index"]]})}
}
\value{
returns an \code{index_list} object, see
  \code{predict_prepare}
}
\description{
Constructs an \code{index_list} from a \code{match_list} and a logical
index for training data
}
\author{
Colman Humphrey
}
