% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_tf}
\alias{is_tf}
\title{Test if \code{x} is a length-one logical}
\usage{
is_tf(x)
}
\arguments{
\item{x}{ideally a logical!}
}
\value{
TRUE if x is a length one logical, else FALSE
}
\description{
Test if \code{x} is a length-one logical
}
\author{
Colman Humphrey
}
\keyword{internal}
