% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahal.R
\name{mahal_imbalance}
\alias{mahal_imbalance}
\title{Computes the Mahalanobis Imbalance of a match, or set of matches}
\usage{
mahal_imbalance(
  x_mat,
  cov_x,
  weight_vec = NULL,
  sqrt_mahal = TRUE,
  match_list = NULL,
  match_list_list = NULL
)
}
\arguments{
\item{x_mat}{numeric matrix (adjust non-numeric columns prior),
already rank-adjusted if desired}

\item{cov_x}{covariance of x, calculated potentially with ranks}

\item{weight_vec}{vector of weights corresponding to columns of x_mat,
giving weights relative to "raw" (ranked) Mahalanobis.
Note that the resulting matrix does depend on the scale of the
weight vector, but the matching won't: scaling the Mahalanobis
matrix has no effect on distance minimising pairs etc}

\item{sqrt_mahal}{logical, default TRUE; do you want regular Mahalanobis:
\eqn{d(x_i, x_j) = (x_i - x_j)' \Sigma^{-1} (x_i - x_j)}
or the square root? (in weighted, \eqn{\Sigma^{-1}} becomes
\eqn{W \Sigma^{-1} W} in the above)}

\item{match_list}{Result of any of the matching methods; provide
exactly one of this param or \code{match_list_list}}

\item{match_list_list}{list of match_list results; provide
exactly one of this param or \code{match_list}}
}
\value{
a vector of mahalanobis distances
}
\description{
This function computes the average distance between treated
and control (not the average absolute or anything), taking ranks into account
if given, and scales this by both the covariances, and the given weights
}
\details{
You can provide a single match, or a list of matches
}
