% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahal.R
\name{mahal_imbalance_list}
\alias{mahal_imbalance_list}
\title{Computed the mahal imbalance for a given match}
\usage{
mahal_imbalance_list(x_mat, match_list, chol_mat, weight_vec, sqrt_mahal)
}
\arguments{
\item{x_mat}{numeric matrix (adjust non-numeric columns prior),
already rank-adjusted if desired}

\item{match_list}{Result of any of the matching methods; provide
exactly one of this param or `match_list_list`}

\item{chol_mat}{}

\item{weight_vec}{vector of weights corresponding to columns of x_mat,
giving weights relative to "raw" (ranked) Mahalanobis.
Note that the resulting matrix does depend on the scale of the
weight vector, but the matching won't: scaling the Mahalanobis
matrix has no effect on distance minimising pairs etc}

\item{sqrt_mahal}{logical, default TRUE; do you want regular Mahalanobis:
\eqn{d(x_i, x_j) = (x_i - x_j)' \Sigma^{-1} (x_i - x_j)}
or the square root? (in weighted, \eqn{\Sigma^{-1}} becomes
\eqn{W \Sigma^{-1} W} in the above)}
}
\description{
Computed the mahal imbalance for a given match
}
\keyword{internal}
