% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_eval.R
\name{match_estimate_tolerance}
\alias{match_estimate_tolerance}
\title{Calculates the "match estimate" for nonbipartite matches:
change per "unit" of tolerance vector}
\usage{
match_estimate_tolerance(
  match_list,
  y_vector,
  tolerance_list = gen_tolerance_list(),
  use_regression = TRUE
)
}
\arguments{
\item{match_list}{Typical \code{match_list} object from
\code{bipartite_matches}.}

\item{y_vector}{The outcome vector.}

\item{tolerance_list}{Typical tolerance list, see
\code{gen_tolerance_list}}

\item{use_regression}{Boolean, use the "naive" average,
or use regression? Default \code{TRUE}}
}
\value{
Returns a single number, the match estimate
}
\description{
This function computes either:
\deqn{
    \frac{1}{m} \sum_1^m \frac{y_{t_i} - y_{c_i}}{t_{t_i} - t_{c_i}}
}
Where \eqn{t} is the tolerance vector, and
\eqn{t_i, c_i} are the \eqn{i}th treatment and control
index.
Or we run a regression:
\deqn{
    y_{t_i} - y_{c_i} ~ t_{t_i} - t_{c_i}
}
And return the coefficient of the tolerance difference
}
\author{
Colman Humphrey
}
