% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_predict.R
\name{match_predict_linear}
\alias{match_predict_linear}
\title{Function to predict treatment / control pairs using
\code{glm} (binomial) or \code{lm}}
\usage{
match_predict_linear(train_test_list, use_lm_approx = FALSE)
}
\arguments{
\item{train_test_list}{output from \code{predict_prepare}}

\item{use_lm_approx}{logical, default FALSE; if}
}
\value{
returns a vector of predictions for the test data
}
\description{
This function takes in training and test data, trains a
"linear" model on the training, predicts on the test, and
returns the test vector. NOTE: the glm model will
fail if the data is too "tricky", so be sure to check.
Also the pure \code{lm} model will be fast
}
\author{
Colman Humphrey
}
