% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_predict.R
\name{match_predict_xgb}
\alias{match_predict_xgb}
\title{Function to predict treatment / control pairs using
xgboost}
\usage{
match_predict_xgb(
  train_test_list,
  nrounds = 50,
  nthread = 1,
  params = list(eta = 0.1, max.depth = 4),
  ...
)
}
\arguments{
\item{train_test_list}{output from \code{predict_prepare}}

\item{nrounds}{training rounds for the xgb algorithm}

\item{nthread}{threads to use for fitting, default 1...}

\item{params}{list of params to pass to xgboost,
most likely something like \code{eta} and \code{max.depth}}
}
\value{
returns a vector of predictions for the test data
}
\description{
This function takes in training and test data, trains an
xgboost model on the training, predicts on the test, and
returns the test vector
}
\author{
Colman Humphrey
}
