% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{match_propensity_list}
\alias{match_propensity_list}
\title{Generates the propensity parameters used for using propensity-based calipers}
\usage{
match_propensity_list(
  propensity_function = propensity_score_xgb(),
  oos_propensity = FALSE,
  n_folds = NULL,
  caliper_sd_mult = 0.6,
  continuous_mult = 100
)
}
\arguments{
\item{propensity_function}{A function that accepts a list
with four elements: \code{x_train}, \code{x_test},
\code{y_train}, \code{y_test}, and forms predictions
using \code{x_test} (I guess \code{y_test} isn't used yet)}

\item{oos_propensity}{Logical, do you want to predict out of sample
for the propensity score? Most people don't, and indeed \code{FALSE}
is the default.}

\item{n_folds}{Default NULL; how many folds you want if using
out of sample propensity.}

\item{caliper_sd_mult}{We'll set the maximum gap between units
as \code{sd(propensity_score) * k}, where this parameter is the
value k. Default 0.6.}

\item{continuous_mult}{See e.g. \code{gen_caliper_list}: instead of
blocking matches that are "too far apart" on the caliper, we'll
add a penalty for going above.}
}
\value{
list with names equal to all input params
}
\description{
We use this for input for \code{all_propensity_caliper_matches} (and
likely a nonbipartite version soon).
}
