% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_standard_dev.R
\name{matrix_to_simple_match}
\alias{matrix_to_simple_match}
\title{Simple function for \code{tol_random_sample} to convert matrices
of pairs into list}
\usage{
matrix_to_simple_match(pairs_matrix, treat_first = TRUE)
}
\arguments{
\item{pairs_matrix}{Matrix with two columns, one for the treatment
indices, one for the control}

\item{treat_first}{Boolean; is the first column the treatment column?}
}
\value{
See \code{tol_random_sample}
}
\description{
Simple function for \code{tol_random_sample} to convert matrices
of pairs into list
}
\author{
Colman Humphrey
}
\keyword{internal}
