% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{min_blocked_rank}
\alias{min_blocked_rank}
\title{Minimum ranked element, not equal to the blocked index}
\usage{
min_blocked_rank(dist_mat, blocked_ind = NULL)
}
\arguments{
\item{dist_mat}{Some distance matrix.}

\item{blocked_ind}{For each row, blocked element.}
}
\value{
Vector of indices
}
\description{
Minimum ranked element, not equal to the blocked index
}
\author{
Colman Humphrey
}
\keyword{internal}
