% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{min_different_rank}
\alias{min_different_rank}
\title{Minimum ranked element, not equal to the row index}
\usage{
min_different_rank(dist_mat)
}
\arguments{
\item{dist_mat}{(Probably symmetric) matrix.}
}
\value{
vector of indices
}
\description{
Minimum ranked element, not equal to the row index
}
\author{
Colman Humphrey
}
\keyword{internal}
