% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{n_sink_generator}
\alias{n_sink_generator}
\title{Generates a function that generates a vector of sink lengths}
\usage{
n_sink_generator(start_frac = 0, end_frac = 0.8, length_out = 9)
}
\arguments{
\item{start_frac}{Smaller fraction of units to use as sink number, default 0.}

\item{end_frac}{Larger fraction of units to use as sink number, default 0.8.}

\item{length_out}{How many sink values we want, default 9.}
}
\value{
Function that accepts a \code{treat_vec}
and returns a vector of numbers.
}
\description{
This function returns a function that accepts a treatment vector
and generates a vector of numbers to be used as sink counts.
}
\author{
Colman Humphrey
}
