% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{near_given_match}
\alias{near_given_match}
\title{Find nearest index to vector given set of indices}
\usage{
near_given_match(find_vec, given_index)
}
\arguments{
\item{find_vec}{vector to find close matches in}

\item{given_index}{index for which we need to find the friends}
}
\value{
set of indices of values closest to each given index
}
\description{
Find nearest index to vector given set of indices
}
\author{
Colman Humphrey
}
