% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_standard_dev.R
\name{nonbipartite_match_sd_scaled}
\alias{nonbipartite_match_sd_scaled}
\title{Simlar to bipartite_match_sd, except for use with
non-bipartite matches.}
\usage{
nonbipartite_match_sd_scaled(
  x_mat,
  cov_x,
  y_vector,
  match_list,
  tolerance_list = gen_tolerance_list(),
  caliper_list = gen_caliper_list(),
  weight_vec = NULL,
  sqrt_mahal = TRUE
)
}
\arguments{
\item{x_mat}{input/design matrix (already rank-adjusted etc)}

\item{cov_x}{the (potentially rank-adjusted) covariance of \code{x_mat}.
This means it's possible that \code{cov(x_mat)} is not equal to \code{cov_x};
see \code{covariance_with_ranks} for more details.}

\item{y_vector}{Outcome vector (not used in match generation).x}

\item{match_list}{A particular match result.}

\item{caliper_list}{Optional, see \code{gen_caliper_list}. Provide
this to force matches that are close on some metric.}

\item{tolerance_vec}{the tolerance used to form the
non-bipartite matches}

\item{tolerance_min}{min value we must obey in tol diffs}

\item{tolerance_max}{max value we must obey in tol diffs}
}
\value{
a single float, the standard deviation (not standard error)
  for the match
}
\description{
For when your result isn't just based
on the differences in the y-vector, but those differences are
weighed inversely proportional to the difference between the
tolerance values
}
\author{
Colman Humphrey
}
