% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite_matches.R
\name{optimal_match}
\alias{optimal_match}
\title{Computes optimal matches, bipartite}
\usage{
optimal_match(dist_mat, treat_vec, n_sinks = 0, tol_val = 0.0001)
}
\arguments{
\item{dist_mat}{Matrix of pairwise distances.}

\item{treat_vec}{Vector representing all subjects; 0 for control,
1 for treated.}

\item{n_sinks}{single value}

\item{tol_val}{tolerance for solving optimal matches - how far is
is acceptable to be from the true optimal value? Speed with large value,
accuracy with small. Only relevant for \code{!with_replacement && !greedy}.
Default 1e-4 is reasonable in many cases.}
}
\description{
Computes optimal matches, bipartite
}
\keyword{internal}
