% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_matches.R
\name{optimal_sink_wrap}
\alias{optimal_sink_wrap}
\title{Given a vector of sink values, generates an optimal match
for each.}
\usage{
optimal_sink_wrap(dist_mat, treat_vec, n_sinks, tol_val)
}
\arguments{
\item{dist_mat}{Matrix of pairwise distances.}

\item{treat_vec}{Vector representing all subjects; 0 for control,
1 for treated.}

\item{n_sinks}{default NULL, vector of sink values to use.}

\item{tol_val}{tolerance for solving optimal matches - how far is
is acceptable to be from the true optimal value? Speed with large value,
accuracy with small. Only relevant for \code{!with_replacement && !greedy}.
Default 1e-4 is reasonable in many cases.}
}
\value{
list of lists; see parent function
}
\description{
Will be slow; you can't just generate one match and subset from it.
}
\author{
Colman Humphrey
}
\keyword{internal}
