% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pair_indexing}
\alias{pair_indexing}
\title{Builds symmetric pairs of indices for vectors}
\usage{
pair_indexing(index_list)
}
\arguments{
\item{index_list}{List of vectors, interpreted as index locations}
}
\value{
Matrix of pairs
}
\description{
For a given vector interpreted as indices, e.g.
\code{c(5, 10, 13)}, we want all possible pairs that
aren't the same element, e.g. we want \code{c(5, 10)},
\code{c(5, 13)} and \code{c(10, 13)}. Further, we want
them in both orders, and then combined into a matrix. Finally,
we want to do this for multiple vectors from a list, and have a
single final matrix output.
}
\author{
Colman Humphrey
}
\keyword{internal}
