% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{paper_mean_functions}
\alias{paper_mean_functions}
\title{Generates the four different mean functions we use
for simulations in the paper.}
\usage{
paper_mean_functions()
}
\value{
List of four functions that take in a matrix and
  output a vector of means:
\describe{
  \item{\code{constant_mu}}{Vector of zeroes}
  \item{\code{linear_mu}}{The matrix multiplied by a random vector.}
  \item{\code{sign_mu}}{Literally \code{sign(x_mat[, 1])}}
  \item{\code{non_linear_mu}}{Messy non-linear function of the data matrix}
}
}
\description{
We make them all mean zero for no real reason, we don't really care
what the mean is.
}
\author{
Colman Humphrey
}
