% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{paper_treatment_functions}
\alias{paper_treatment_functions}
\title{Generates the four different treatment functions we use
for simulations in the paper.}
\usage{
paper_treatment_functions(target_mean = 0.425)
}
\arguments{
\item{target_mean}{Desired mean for each of the treatment probabilities,
should be in (0, 1).}
}
\value{
List of four functions that take in a matrix and
  output a vector of probabilities:
\describe{
  \item{\code{constant_treat_prob}}{Just repeats \code{target_mean}}
  \item{\code{logistic_treat_prob}}{Computes a logistic style relationship
  between the data matrix and the output vector}
  \item{\code{sparse_treat_prob}}{Based on the sign of the first data matrix
  column; gives a number a bit above the mean and a bit below.}
  \item{\code{sparse_nonlin_treat_prob}}{Cubic logit function of the
  first vector}
}
}
\description{
Generates the four different treatment functions we use
for simulations in the paper.
}
\author{
Colman Humphrey
}
