% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{parallel_sim}
\alias{parallel_sim}
\title{Run \code{compute_sim_result} in parallel using \code{parallel::mclapply}}
\usage{
parallel_sim(
  x_generator = default_x_generator,
  treat_prob_generator,
  mean_generator,
  error_generator = default_error_generator,
  n_sink_gen = n_sink_generator(),
  match_method = "with_replacement",
  n_rows = 500L,
  n_cols = 5L,
  num_weight_vectors = 100L,
  num_cores = parallel::detectCores() - 1,
  iterations = 100L,
  names_list = NULL,
  silent = !interactive()
)
}
\arguments{
\item{x_generator}{Function that takes number of rows and columns
and produces a data matrix of that dimension.}

\item{treat_prob_generator}{Function that takes a matrix and produces
treatment probabilities.}

\item{mean_generator}{Function that takes a matrix and produces
an expected value for each row.}

\item{error_generator}{Function that accepts a number of rows and
generates an error vector (e.g. generates normal noise). Should
generate zero-mean data.}

\item{n_sink_gen}{Default \code{n_sink_generator}, and you'll
probably want to use that: this argument should be a function that
accepts a \code{treat_vec} and produces a vector of sink numbers.}

\item{n_rows}{How many rows to generate.}

\item{n_cols}{How many columns to use.}

\item{num_weight_vectors}{How many weight vectors to generate.}

\item{num_cores}{How many cores to use.}

\item{iterations}{How many iterations to do.}

\item{silent}{Default \code{!interactive()}, if you want to
suppress messages.}
}
\description{
Run \code{compute_sim_result} in parallel using \code{parallel::mclapply}
}
\author{
Colman Humphrey
}
