% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_eval.R
\name{permutation_brier}
\alias{permutation_brier}
\title{For a given match, computes the brier score distribution
if the pairing were truly random}
\usage{
permutation_brier(
  x_mat,
  match_list,
  design = "cross_all",
  use_cv = TRUE,
  num_permutations = 100L,
  match_predict_function = match_predict_xgb(),
  num_folds = 5,
  train_fraction = 0.7
)
}
\arguments{
\item{x_mat}{typical input matrix}

\item{match_list}{match result}

\item{design}{see \code{predict_prepare}}

\item{use_cv}{logical, default TRUE: use CV to get briers? Else
split.}

\item{num_permutations}{how many permutations to do}

\item{match_predict_function}{function to predict treated units}

\item{num_folds}{if using CV, how many folds?}

\item{train_fraction}{if using split, fraction to train?}
}
\value{
vector of brier scores for random pairings
}
\description{
Works for bipartite and non-bipartite, since it just
switches the labels within pairs.
}
\author{
Colman Humphrey
}
