% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_weighted_matches.R
\name{permutation_matches}
\alias{permutation_matches}
\title{Takes matches and their Brier scores, and computes
permutation Brier scores and the best matches}
\usage{
permutation_matches(
  matches_by_sinks,
  briers_by_sinks,
  x_mat,
  n_sinks,
  approximate_by_best = TRUE,
  silent = !interactive()
)
}
\arguments{
\item{matches_by_sinks}{List by number of sinks, each a list
of match results (a match list), for each weight vector.}

\item{briers_by_sinks}{List by number of sinks, each a vector
of Brier results. Basically a number for each match in
\code{matches_by_sinks}.}

\item{x_mat}{Typical input matrix}

\item{n_sinks}{Vector of number of sinks - probably could get this
directly from \code{matches_by_sinks}, but nice to be explicit.}

\item{approximate_by_best}{Logical, default \code{TRUE}. Only
compute one permutation distribution, using the best result by brier
score to do so. Useful because it changes little, but saves a ton of
time.}

\item{silent}{Do you want to suppress message output? Default
\code{!interactive()}.}
}
\value{
Returns a list of two lists. The first is vectors
  of permutation Brier scores (one per match). The second is the
  best match at each sink value, along with some extra info about that
  match.
}
\description{
Can work for bipartite or non-bipartite matches,
the permutation is just over the labels.
}
\author{
Colman Humphrey
}
