% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_plots.R
\name{plot_sims}
\alias{plot_sims}
\title{Plotting sim results}
\usage{
plot_sims(all_results, cut_y = 0.8)
}
\description{
long_results_list should have length four:
n = 500, 1000, 2000, 4000
each should then have length four:
p = 5, 10, 20, 40
each should then have length 16:
models (a, b, c, d) vs (a, b, c, d)
}
\details{
plan: four plots, one for each N value
}
