% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{propensity_bipartite_matches}
\alias{propensity_bipartite_matches}
\title{Propensity match for (default) bipartite}
\usage{
propensity_bipartite_matches(
  x_mat,
  treat_vec,
  match_method = c("with_replacement", "optimal", "greedy"),
  propensity_list = gen_propensity_list(),
  n_sinks = 0,
  caliper_list = gen_caliper_list(),
  sqrt_mahal = TRUE,
  tol_val = NULL
)
}
\arguments{
\item{x_mat}{input/design matrix (already rank-adjusted etc)}

\item{treat_vec}{Logical (or 1/0) vector, indicating treatment (or control).}

\item{propensity_list}{See \code{gen_propensity_list}}

\item{n_sinks}{how many potential matches to not bother with
NOTE: you can do this as a vector, but not for optimal matching.}

\item{caliper_list}{Optional, see \code{gen_caliper_list}. Provide
this to force matches that are close on some metric.}

\item{tol_val}{For optimal matches, you can set a tolerance
to be within optimality of, which can be zero for perfect optimality.
Default 1e-4 is reasonable in many cases.}
}
\description{
Propensity match for (default) bipartite
}
