% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{propensity_score}
\alias{propensity_score}
\title{Calculates propensity scores for a given matrix and treatment vector}
\usage{
propensity_score(x_mat, treat_vec, propensity_list = gen_propensity_list())
}
\arguments{
\item{x_mat}{Standard input matrix (already rank adjusted).}

\item{treat_vec}{Usual 0/1 treatment vector.}

\item{propensity_list}{See \code{gen_propensity_list}}
}
\value{
Returns a vector equal in length to treat_vec of propensity.
  score.
}
\description{
This function takes in an input matrix and a treatment vector,
along with a function that makes predictions (default xgboost method given)
and returns a predicted probability of treatment for each unit, either
using in-sample or out-of-sample fits.
}
\author{
Colman Humphrey
}
