% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{propensity_score_linear}
\alias{propensity_score_linear}
\title{Function factory to predict treatment using \code{glm} (binomial)
or \code{lm}}
\usage{
propensity_score_linear(use_linear_lm = FALSE)
}
\value{
returns a function that accepts \code{train_test_list}
and this returns a vector of predictions for the test data
}
\description{
Does simple wrap around \code{glm} / \code{lm}
The returned function accepts one parameter, \code{train_test_list},
a list with \code{x_train}, \code{x_test}, \code{y_train}, \code{y_test}
}
\author{
Colman Humphrey
}
