% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rank_integer_index}
\alias{rank_integer_index}
\title{Converts \code{rank_cols} in all allowed forms to an integer index}
\usage{
rank_integer_index(rank_cols, x_mat)
}
\arguments{
\item{rank_cols}{integer/number, or logical, or names within
\code{colnames(x_mat)}}

\item{x_mat}{the x matrix of interest}
}
\description{
Takes in ranked cols, either given by integer (numeric is fine) index;
logical index; or named. Converts to integer index
}
\author{
Colman Humphrey
}
\keyword{internal}
