% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ranked_x}
\alias{ranked_x}
\title{Converts indicated columns to ranked versions of themselves}
\usage{
ranked_x(x_mat, rank_cols = NULL, ties_method = "average")
}
\arguments{
\item{x_mat}{numeric matrix (adjust non-numeric columns prior)}

\item{rank_cols}{names or index of columns to be converted to ranks
before analysis}

\item{ties_method}{how to break ties in ranks, by default uses
"average" (same as the rank function's default). See \code{?rank}
for further options}
}
\value{
x_mat again, potentially adjusted for ranks
}
\description{
This function takes a numeric matrix and converts any columns
indicated by the \code{rank_cols} input to their ranks
(break ties however you want), scaled down by \code{nrow(x_mat)}.
}
\author{
Colman Humphrey
}
\keyword{internal}
