% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_eval.R
\name{regression_eval}
\alias{regression_eval}
\title{Evaluate match using regression}
\usage{
regression_eval(match_list, y_vector, tolerance_list = gen_tolerance_list())
}
\arguments{
\item{match_list}{Usual match list}

\item{y_vector}{Usual outcome vector}

\item{tolerance_list}{Usual tolerance list}
}
\value{
Returns a list:
\itemize{
\item \code{estimate}: the estimated treatment effect. If we're in bipartite
land, then this is the usual difference in means, else
it's the effect of a "unit" change in tolerance
\item \code{standard_error}: the standard deviation of the above coefficient
}
}
\description{
Generates a list with the estimated treatment effect,
and the standard error for that effect. If bipartite
(i.e. no tolerance used) just regresses differences in
\code{y_vector} against the intercept, basically difference in means.
If a tolerance vector is used, the differences are regressed
against the tolerance differences.
In both cases, if controls are repeated, we build up the appropriate
correlation matrix and use Generalised Linear Regression to solve.
}
\author{
Colman Humphrey
}
