% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonbipartite_matches.R
\name{reorder_nbp}
\alias{reorder_nbp}
\title{Reorders list first by treat index, then within pairs by
highest tol val if given}
\usage{
reorder_nbp(match_list, tolerance_vec = NULL)
}
\description{
Reorders list first by treat index, then within pairs by
highest tol val if given
}
\keyword{internal}
