% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{reshape_p_cut_list}
\alias{reshape_p_cut_list}
\title{Reshapes a list of simulations to a nice dataframe, by p-cut}
\usage{
reshape_p_cut_list(
  list_of_sims,
  treat_model_name,
  mu_model_name,
  p_cut,
  n_rows,
  n_cols,
  num_weight_vectors
)
}
\arguments{
\item{list_of_sims}{List of results from \code{compute_sim_results}.}

\item{treat_model_name}{Name of the treatment model.}

\item{mu_model_name}{Name of the mean generation model.}

\item{p_cut}{Vector of one-sided permution p-values to use.}

\item{n_rows}{How many rows were used.}

\item{n_cols}{How many columns were used.}

\item{num_weight_vectors}{How many weight vectors were used}
}
\value{
Data frame spreading results...
}
\description{
Reshapes a list of simulations to a nice dataframe, by p-cut
}
\author{
Colman Humphrey
}
