% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{swap_pairs}
\alias{swap_pairs}
\title{Given a logical swap vector, switches treated and control units around}
\usage{
swap_pairs(match_list, swap)
}
\arguments{
\item{match_list}{typical \code{match_list} entity}

\item{swap}{logical vector indicating which pairs should swap}
}
\value{
another match_list, but now with swapped pairs
}
\description{
Given a logical swap vector, switches treated and control units around
}
\author{
Colman Humphrey
}
