% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sym_mat}
\alias{sym_mat}
\title{Symmetrises a matrix in a boring way, and zeros the diag}
\usage{
sym_mat(mat)
}
\arguments{
\item{mat}{a matrix}
}
\description{
This is useful for generating pairwise distance matrices
}
\keyword{internal}
