% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_utils.R
\name{target_mean_expit}
\alias{target_mean_expit}
\title{Shifts the linear component until the mean of the
expit is the target mean.}
\usage{
target_mean_expit(target_mean, linear_vector)
}
\arguments{
\item{target_mean}{Desired target mean.}

\item{linear_vector}{Basically \eqn{X \beta} (no shift).}
}
\value{
\code{expit(linear_vector + a)} for some \code{a} such that
  \code{mean(expit(linear_vector + a))} is close to \code{target_mean}.
}
\description{
Shifts the linear component until the mean of the
expit is the target mean.
}
\author{
Colman Humphrey
}
