% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_standard_dev.R
\name{tol_random_sample}
\alias{tol_random_sample}
\title{Generates a random pairing from a tolerance list in a naive way}
\usage{
tol_random_sample(
  tolerance_list,
  prior_pairs = matrix(NA, nrow = 0L, ncol = 2L),
  pairable_units = NULL,
  iteration = 0L,
  max_iterations = 10L,
  verbose = FALSE
)
}
\arguments{
\item{tolerance_list}{Usual tol, see e.g. \code{gen_tolerance_list}}

\item{prior_pairs}{For use in recursion - what pairs are already formed
(indeed we have none for the first iteration)}

\item{pairable_units}{For use in recursion - what units are available
to form pairs (\code{NULL} is all, hence for the first iteration)}

\item{iteration}{What iteration number are we on}

\item{max_iterations}{Max iterations}

\item{verbose}{Boolean - spit out extra messages / warnings?}
}
\value{
"Simple" match list, list with two elements:
\itemize{
   \item{\code{treat_index}} Treatment units
   \item{\code{control_index}} Control units
}
}
\description{
This function takes in a tolerance list and attempts to generate
a "simple" match list by naive random sampling.
}
\author{
Colman Humphrey
}
\keyword{internal}
