% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tolerance_check}
\alias{tolerance_check}
\title{Validates a match on a tolerance list}
\usage{
tolerance_check(match_list, tolerance_list)
}
\arguments{
\item{match_list}{Typical \code{match_list} object from
\code{bipartite_matches}.}

\item{tolerance_list}{Typical tolerance list, see
\code{gen_tolerance_list}}
}
\value{
List with two elements:
\itemize{
 \item{\code{error}}{Boolean - is all good?}
 \item{\code{message}}{If error, then gives error string}
}
}
\description{
Checks that the differences are all at least
at the min, and if a max is given, checks that too.
Gives reasonable error messages on errors
}
\author{
Colman Humphrey
}
\keyword{internal}
