% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonbipartite_matches.R
\name{tolerance_to_caliper_list}
\alias{tolerance_to_caliper_list}
\title{Converts tolerance list to caliper list}
\usage{
tolerance_to_caliper_list(tolerance_list, use_min = TRUE, continuous_mult = 1)
}
\arguments{
\item{tolerance_list}{Result from \code{gen_tolerance_list}}

\item{use_min}{Logical, should the caliper max be the tolerance min?
Use tolerance max if not. Default \code{TRUE}.}

\item{continuous_mult}{The value to multiply differences above caliper max.
Set as \code{Inf} to have infinite penalties, i.e. block matches above
the max.}
}
\value{
List from \code{gen_caliper_list}
}
\description{
Converts tolerance list to caliper list
}
\author{
Colman Humphrey
}
\keyword{internal}
