% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_size_sub}
\alias{unique_size_sub}
\title{Takes a subsample with a given number of unique elements}
\usage{
unique_size_sub(primary, unique)
}
\arguments{
\item{primary}{vector to subsample from}

\item{unique}{how many unique elements you want}
}
\value{
vector equal in length to primary, with `unique` unique elements
}
\description{
Takes a subsample with a given number of unique elements
}
\author{
Colman Humphrey
}
\keyword{internal}
