% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_utils.R
\name{x_from_cov}
\alias{x_from_cov}
\title{Generates a matrix where each row is generated
with covariance equal to \code{cov_mat}.}
\usage{
x_from_cov(cov_mat, n_rows)
}
\arguments{
\item{cov_mat}{The covariance desired.}

\item{n_rows}{How many rows to produce.}
}
\value{
A matrix such that the covariance would tend to
  \code{cov_mat} as \code{n_rows} grows.
}
\description{
Generates a matrix where each row is generated
with covariance equal to \code{cov_mat}.
}
\author{
Colman Humphrey
}
