% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_standard_dev.R
\name{y_tolerance_diff_ratio}
\alias{y_tolerance_diff_ratio}
\title{Quick and dirty method to approximate the sd and mean
of the ratio of the differences}
\usage{
y_tolerance_diff_ratio(y_vector, tolerance_list, samples = NULL)
}
\arguments{
\item{y_vector}{Vector of outcomes we care about}

\item{tolerance_list}{Usual tolerance list (see \code{gen_tolerance_list})}

\item{samples}{How many samples to use. Defaults to a number between 50
and 200, depending on the length of the various vectors}
}
\value{
Returns a single number: the average of the sds of \code{samples}
random runs
}
\description{
For non-bipartite work, we work with something like:
\deqn{
    \frac{y_{\text{treat}} - y_{\text{control}}}
         {tol_{\text{treat}} - tol_{\text{control}}}
}
And not just \eqn{y_{\text{treat}} - y_{\text{control}}} as
our "difference": here we say we're caring about the change
in \eqn{y} per unit tol. Note that we use regression
to estimate, but that is really just a statment about error
assumptions.
This function estimates the standard deviation of this
if you just did random pairings \strong{that obeyed the tolerance
rules} with min and max.
This function computes approx sd for the \strong{optimal} case.
Also yes this is a terrible function name.
}
\author{
Colman Humphrey
}
