% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_plots.R
\name{plot_sims}
\alias{plot_sims}
\title{Plotting simulation results. Plan: four plots, one for each N value}
\usage{
plot_sims(all_results, y_rel = c(0.1, 0.4))
}
\arguments{
\item{all_results}{result from \code{generate_plot_list}}

\item{y_rel}{where to limit y to above and below (length two vector)}
}
\value{
Nothing, just plots to the device
}
\description{
Plotting simulation results. Plan: four plots, one for each N value
}
\author{
Colman Humphrey
}
