% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{propensity_score_xgb}
\alias{propensity_score_xgb}
\title{Function factory to predict treatment using xgboost}
\usage{
propensity_score_xgb(
  nrounds = 50,
  nthread = 1,
  params = list(eta = 0.1, max.depth = 3),
  ...
)
}
\arguments{
\item{nrounds}{training rounds for the xgb algorithm}

\item{nthread}{threads to use for fitting, default 1...}

\item{params}{list of params to pass to xgboost,
most likely something like \code{eta} and \code{max.depth}}
}
\value{
returns a function that accepts \code{train_test_list}
and this returns a vector of predictions for the test data
}
\description{
This operates the exact same as \code{match_predict_xgb}, which
is itself quite a simple wrap of regular xgboost code.
Main difference here is the input data is even simpler.
The returned function accepts one parameter, \code{train_test_list},
a list with \code{x_train}, \code{x_test}, \code{y_train}, \code{y_test}
}
\author{
Colman Humphrey
}
